functions/requirements.txt
google-cloud-storage==1.10.0
jsonschema==2.6.0
